+++ src/wise/dataloader/dataset.py
@@ -615,7 +615,7 @@ def _get_dataset(
     offset: float | None = None,
     thumbnails: bool = True,
 ) -> Optional[MediaDataset]:
-    if media_type == MediaType.AV:
+    if media_type is MediaType.AV:
         if video_frames_per_chunk <= 0 and audio_samples_per_chunk <= 0:
             logger.warning(
                 "Both video_frames_per_chunk and audio_samples_per_chunk are <= 0, skipping video files"
@@ -632,7 +632,7 @@ def _get_dataset(
             offset=offset,
             thumbnails=thumbnails,
         )
-    elif media_type == MediaType.VIDEO:
+    elif media_type is MediaType.VIDEO:
         if video_frames_per_chunk <= 0:
             logger.warning(
                 "video_frames_per_chunk is <= 0, skipping video-only files"
@@ -646,7 +646,7 @@ def _get_dataset(
             offset=offset,
             thumbnails=thumbnails,
         )
-    elif media_type == MediaType.AUDIO:
+    elif media_type is MediaType.AUDIO:
         if audio_samples_per_chunk <= 0:
             logger.warning(
                 "audio_samples_per_chunk is <= 0, skipping audio-only files"
@@ -659,7 +659,7 @@ def _get_dataset(
             preprocessing_function=audio_preprocessing_function_map,
             offset=offset,
         )
-    elif media_type == MediaType.IMAGE:
+    elif media_type is MediaType.IMAGE:
         stream = ImageDataset(
             input_files,
             preprocessing_function=image_preprocessing_function_map,
@@ -672,13 +672,13 @@ def _get_dataset(
 
 def get_dataset(media_metadata: list[DatasetPayload], params: dict[str, Any]):
     # sort and group (by media_type - image/video/audio/av)
-    sort_func = lambda x: x.media_type
+    sort_func = lambda x: x.media_type.name
     sorted_metadata = sorted(media_metadata, key=sort_func)
     datasets = list(
         filter(
             None,
             (
-                _get_dataset({x.id: x.path for x in g}, k, **params)
+                _get_dataset({x.id: x.path for x in g}, MediaType[k], **params)
                 for k, g in itertools.groupby(sorted_metadata, key=sort_func)
             ),
         )